Fix compression dictionary clang osx error
Summary:
There was one narrowing conversion in D52287 that only showed up with
clang on osx.

Test Plan:
  $ make clean && USE_CLANG=1 DISABLE_JEMALLOC=1 TEST_TMPDIR=/dev/shm/rocksdb OPT=-g make -j32 check

Reviewers: sdong, lightmark, IslamAbdelRahman

Reviewed By: IslamAbdelRahman

Subscribers: andrewkr, dhruba, leveldb

Differential Revision: https://reviews.facebook.net/D57357

+++ db/db_test2.cc
@@ -912,9 +912,9 @@ TEST_F(DBTest2, PresetCompressionDict) {
       std::vector<std::string> files;
       GetSstFiles(dbname_, &files);
       for (const auto& file : files) {
-        size_t curr_bytes;
+        uint64_t curr_bytes;
         env_->GetFileSize(dbname_ + "/" + file, &curr_bytes);
-        out_bytes += curr_bytes;
+        out_bytes += static_cast<size_t>(curr_bytes);
       }
 
       for (size_t j = 0; j < kNumL0Files * (kL0FileBytes / kBlockSizeBytes);
